Switched to ScienceFilePath, moved fuctions over to cdf_utils, used loggers and fixed some return things.
